Resolved persistence issue with league level coaches & managers and Persisting GameConfig

+++ src/main/java/com/IceHockeyLeague/LeagueManager/GamePlayConfig/IGameResolverConfig.java
@@ -1,6 +1,6 @@
 package com.IceHockeyLeague.LeagueManager.GamePlayConfig;
 
 public interface IGameResolverConfig {
-    void setRandomWinChance(double winChance);
-    double getRandomWinChance();
+    void setRandomWinChance(float winChance);
+    float getRandomWinChance();
 }
